// apps/desktop/src/common/viewmodel/conversation/main/controller/types.ts
import type {
    MessageType,
    PollAnnounceType,
    PollAnswerType,
    PollDisplayMode,
    PollState,
} from '~/common/enum';
import type {
    OutboundFileMessageInitFragment,
    OutboundImageMessageInitFragment,
    OutboundPollMessageInitFragment,
    OutboundTextMessageInitFragment,
    OutboundVideoMessageInitFragment,
    OutboundAudioMessageInitFragment,
} from '~/common/network/protocol/task/message-processing-helpers';
import type {IdentityString, MessageId, PollId} from '~/common/network/types';
import type {Dimensions, f64, ReadonlyUint8Array, u53} from '~/common/types';
import type {transcodeAudioToMp4Aac, transcodeAudioToMp4Opus} from '~/common/utils/audio';
import type {transcodeVideoToMp4H264} from '~/common/utils/video';

/**
 * Required data the {@link ConversationViewModelController} needs to send a message.
 */
export type SendMessageEventDetail =
    | SendTextBasedMessageInformation
    | SendFileBasedMessageInformation
    | SendPollBasedMessageInformation;

export interface SendTextBasedMessageInformation {
    readonly type: 'text';
    readonly text: string;
    readonly quotedMessageId?: MessageId | undefined;
}

export interface SendFileBasedMessageInformation {
    readonly type: 'files';
    readonly files: {
        readonly bytes: ReadonlyUint8Array;
        readonly thumbnailBytes?: ReadonlyUint8Array;
        readonly caption?: string;
        readonly fileName: string;
        readonly fileSize: u53;
        readonly mediaType: string;
        readonly thumbnailMediaType?: string;
        readonly dimensions?: Dimensions;
        readonly sendAsFile: boolean;
        readonly duration?: u53;
    }[];
}

export interface SendPollBasedMessageInformation {
    readonly type: 'poll';
    readonly description: string;
    readonly answerType: PollAnswerType;
    readonly announceType: PollAnnounceType;
    readonly displayMode: PollDisplayMode;
    readonly choices: {
        readonly choiceId: u53;
        readonly description: string;
    }[];
    readonly pollState: PollState;
}

export interface TextMessageWithByteLength {
    readonly type: 'text';
    readonly text: string;
    readonly byteLength: u53;
}

export interface PollLookup {
    readonly pollCreatorIdentity: IdentityString;
    readonly pollId: PollId;
}

/**
 * Partial data the {@link ConversationViewModelController} needs to prepare a message for sending.
 */
export type OutboundMessageInitFragment =
    | Omit<OutboundTextMessageInitFragment, 'direction' | 'id' | 'createdAt'>
    | Omit<OutboundFileMessageInitFragment, 'direction' | 'id' | 'createdAt'>
    | Omit<OutboundImageMessageInitFragment, 'direction' | 'id' | 'createdAt'>
    | Omit<OutboundVideoMessageInitFragment, 'direction' | 'id' | 'createdAt'>
    | Omit<OutboundAudioMessageInitFragment, 'direction' | 'id' | 'createdAt'>
    | Omit<OutboundPollMessageInitFragment, 'direction' | 'id' | 'createdAt'>;

export type TranscodeFunction =
    | typeof transcodeVideoToMp4H264
    | typeof transcodeAudioToMp4Aac
    | typeof transcodeAudioToMp4Opus;

export interface TranscodingResult {
    readonly type: MessageType.AUDIO | MessageType.FILE | MessageType.VIDEO;
    readonly bytes: ReadonlyUint8Array;
    readonly duration: f64;
    readonly mediaType: string;
    readonly fileName: string;
    readonly fileSize: u53;
}